1. 修改积分分类类型code

+++ src/main/java/com/hds/xquark/dal/mapper/GradeCodeMapper.java
@@ -16,7 +16,7 @@ public interface GradeCodeMapper {
 
   GradeCode selectByFunctionCode(String functionCode);
 
-  GradeCode selectByCategoryIdAndStatus(@Param("categoryId") Long categoryId,
+  GradeCode selectByCategoryIdAndStatus(@Param("categoryId") Integer categoryId,
       @Param("status") Integer status);
 
   List<GradeCode> list(@Param("offset") Integer offset, @Param("size") Integer size);
